+++ sdk/storage/storage-file-datalake/test/node/sas.spec.ts
@@ -3,7 +3,7 @@
 
 import { UserDelegationKey } from "@azure/storage-blob";
 import { record, Recorder } from "@azure-tools/test-recorder";
-import * as assert from "assert";
+import { assert } from "chai";
 import { Context } from "mocha";
 
 import {