+++ sdk/test-utils/recorder-new/src/index.ts
@@ -5,5 +5,4 @@ export { recorderHttpPolicy, RecorderClient } from "./core-v2-recorder";
 export { TestProxyHttpClientCoreV1, RecorderClientCoreV1 } from "./core-v1-recorder";
 export { relativeRecordingsPath } from "./utils/relativePathCalculator";
 export { SanitizerOptions, RecorderStartOptions } from "./utils/utils";
-export { NoOpCredential } from "./recorderAADCredential";
 export { env } from "./utils/env";